requirements.txt
discord.py==2.6.4
yt-dlp==2025.12.8
PyNaCl==1.5.0
imageio-ffmpeg==0.6.0
